fix(gtfs): Update gtfs-lib to v1.4.0
This commit updates gtfs-lib to v1.4.0 to bring it to parity with the version of gtfs-api that is
included. It also updates the style of creating bundleScopedFeedIds (how we reference the stored zip
and db files) from using underscores to hyphens. This will require a database migration and a
filename replacement of all stored feeds.

addresses #56