clear worker startup time once workers have started up
We were recording when workers began starting up so we could avoid repeatedly starting up more of them while waiting for the first one to start up.
Once the worker started up we never cleared out that value, so Analysis would refuse to start up another worker in the same category for one hour.
In code branches where we have evidence that a worker exists, we now clear out the flag saying we're waiting for it to exist.