get incomplete regional analysis results
This is unfinished: there are two different endpoints, one for complete and one for incomplete results, but the main endpoint for complete results is now detecting whether the analysis is complete or incomplete.

This will probably not work because the UI expects a "wrapped URL" for redirecting to finished results on S3, but the unfinished results are streamed back from a file immediately.