feat(broker): mimic content-type headers from worker
This is necessary for conveyal/r5#459 to work properly.
The workers can generate responses in several formats including GeoTIFF, proprietary binary grid, or JSON.
We want to preserve that information when the response is repeated by the broker back to the UI.
It's still not entirely clear how well the gzip compression is being handled - whether it's being done at all, or whether we're inflating and re-deflating on the broker.