fix(AggregationAreaController): convert to WGS84
Previously it was streaming the raw, possibly projected geometries
without even checking whether they were in WGS84 or not. It did not
complain that they were not in WGS84, just said they were too big (since
the units were meters or feet instead of degrees as expected).

This is one small part of #213 which we are applying as a bugfix.