fix(FastRaptorWorker): detect stops updated in a given round
Maintaining sets of updated stops gets complex when layering frequency
routes on top of scheduled results and frequency upper bound results.
instead this is checking on the fly whether a stop's arrival time is
earlier than the one in the previous round. This should be reasonably
efficient since it is accessing arrays that are already in heavy use in
neighboring code, and often accessing those arrays sequentially.