track updated stops separately for scheduled and frequency searches
The scheduled search and upper bound search benefit from tracking which
stops were updated in the current departure minute, rather than in the
previously evaluated departure minutes. In fact, if we don't do this the
range raptor optimization is ineffective.

Frequency routes must look at the combined effects of all departure
minutes - the frequency route's schedule will have changed, but it may
be reached by a ride that was found in a different departure minute.